ZSH: Use nodenv and rbenv again, but without using eval and init
Using `eval "$(rbenv init -)"` adds delay to the boot time of the shell.
The script generated by rbenv adds a line that calls `rbenv rehash`,
which can be very slow. This line can be removed by using the
`--no-rehash` option, but then the `eval` call itself adds also some
delay. The only mandatory step for rbenv to work is the export PATH.

The same applies to nodenv.