neovim: Remove marks plugin
There is a bug in nvim where after deleting local marks they reapear in
the file after closing and opening the editor again. It is something to
do with the global shada file (more info[1]). It can be solved by doing
a `:wshada!`, but I'm too lazy to try to automate that.

[1]: https://github.com/neovim/neovim/issues/4295