neovim: Yet another LSP refactor
The main change is that now LSP servers are configured only on FileType
events, and they don't use lspconfig or mason-lspconfig.

To find if there is a binary to execute the LSP server I added a new
module, binaries. This module will try to find a binary given a certain
rules. In general it will try:

* To find the binary in the current system (PATH)
* To find the binary in Mason, and install it if is not installed
  already

These rules can be changed. For example, rust-analyzer tries to use
rustup to run it, if not it fallbacks to Mason, sorbet only tries to
find a local bin, etc.

Given how Mason installation works, the whole process has to be async,
so a lot of callbacks around.

I have no idea if I'm going to keep this, but it was interesting to play
with it.